pyproject.toml
[project]
name = "listless"
version = "0.1.0"
description = "Radarr/Sonarr-compatible list server with modular providers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",
    "beautifulsoup4>=4.12.0",
    "fastapi>=0.115.0",
    "httpx[http2]>=0.28.1",
    "pydantic-settings>=2.4.0",
    "sqlalchemy>=2.0.0",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "httpx>=0.28.1",
    "coverage>=7.13.4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
